requirements.txt
mcp>=1.0.0
skyfield>=1.49
numpy>=1.26
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.0.0
httpx>=0.27.0
blaxel>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.38.0
plotly>=5.22.0
pandas>=2.2.0
